Guard against linker script when built as a dep
Adds a check for both `CARGO_PRIMARY_PACKAGE` and `CARGO_BIN_NAME` in
the build script before injecting the linker script `memory.x`. This
should prevent it from being included when building this crate as a
dependency.

Without this guard, this crate's `memory.x` will be picked up by the
linker when building an dependant crate, overriding that crate's own
`memory.x`.